+++ src/wfg/ltv_econ/plugins/LtvEconomyModPlugin.java
@@ -43,30 +43,6 @@ public class LtvEconomyModPlugin extends BaseModPlugin {
         if (manager.isModEnabled(LUNA_LIB)) {
             LunaSettings.addSettingsListener(new ConfigLunaSettingsListener());
         }
-
-        // TODO find a way to attach a panel to the title screen. 
-        // boolean showWarnings = false;
-        // final StringBuilder txt = new StringBuilder("Warning!");
-        // if (manager.isModEnabled(GRAND_COL)) {
-        //     showWarnings = true;
-        //     txt.append("\nLTV-Economy already has a scroll bar for industries."+
-        //         " Using Grand.Colonies will break the game."
-        //     );
-        // }
-
-        // if (manager.isModEnabled(ASTRAL_ASCENT)) {
-        //     showWarnings = true;
-        //     txt.append("\nAstral Ascension is not supported by LTV-Economy."+
-        //         " Use at your own risk."
-        //     );
-        // }
-
-        // if (showWarnings) {
-            // final DialogPanel warningPanel = new DialogPanel(
-            //     Attachments.getScreenPanel(), null, txt.toString(), "Close"
-            // );
-            // warningPanel.show(0.3f, 0.3f);
-        // }
     }
 
     @Override